Base the container on ubi8/ubi-init image
This allows us to remove all the hacks we had to do to get systemd
running on a regular centos container.
The majority of the packages that we need though, are not included
in the default ubi repos so we suppliment with the CentOS 8 repos

The other big change here is using the packaged ruby rather than
building it as a part of the image build.
This just seemed easier.