Switch to running as the apache user by default instead of root
In k3s, we were failing to start because our security policy disables running
as root and that is the default user from the base container.

When running in OpenShift, the pod is run with a random UID, groups root & random ID
When running in k3s, the pod is now run as apache:apache

The permissions need to allow writing to both directories in both cases.